// src/app/04_SIR.Exportaciones.Reportes/Interfaces/DataKosherAgrupada.interface.ts
export interface DataKosherAgrupada {
    mercaderia:         string;
    cantidadPallets:    number;
    cantidadCajas:      number;
    kilosNetos:         number;
    kilosBrutos:        number;
    precioTotal:        number;
    precioPromedio:     number;
    especieTipo:        string;
    container:          string;
}